+++ Source/EventFlow/Core/GuidFactories.cs
@@ -29,6 +29,31 @@ namespace EventFlow.Core
     public static class GuidFactories
     {
         /// <summary>
+        /// Creates a sequential Guid that can be used to avoid database fragmentation
+        /// http://stackoverflow.com/a/2187898
+        /// </summary>
+        public static class Comb
+        {
+            public static Guid Create()
+            {
+                var destinationArray = Guid.NewGuid().ToByteArray();
+                var time = new DateTime(0x76c, 1, 1);
+                var now = DateTime.Now;
+                var span = new TimeSpan(now.Ticks - time.Ticks);
+                var timeOfDay = now.TimeOfDay;
+                var bytes = BitConverter.GetBytes(span.Days);
+                var array = BitConverter.GetBytes((long)(timeOfDay.TotalMilliseconds / 3.333333));
+
+                Array.Reverse(bytes);
+                Array.Reverse(array);
+                Array.Copy(bytes, bytes.Length - 2, destinationArray, destinationArray.Length - 6, 2);
+                Array.Copy(array, array.Length - 4, destinationArray, destinationArray.Length - 4, 4);
+
+                return new Guid(destinationArray);
+            }
+        }
+
+        /// <summary>
         /// Creates a name-based UUID using the algorithm from RFC 4122 §4.3.
         /// http://code.logos.com/blog/2011/04/generating_a_deterministic_guid.html
         /// </summary>