+++ RELEASE_NOTES.md
@@ -1,5 +1,7 @@
 ### New in 0.21 (not released yet)
 
+* New: Added `Identity<>.NewComb()` that creates sequential unique IDs that can
+  be used to minimize database fragmentation
 * New: Added `IReadModelContext.Resolver` to allow read models to fetch
   additional resources when events are applied